requirements.txt
django 
djangorestframework
psycopg2 
redis 
djangorestframework-simplejwt